Kill interpreter dispatch loop, use handleResume() + dispatchBB()
Summary:
We rarely use interpreter for the whole request. When we do, it is clearly not
performance sensitive, otherwise we would have used JIT instead.

Instead of using dispatch(), route the logic through handleResume(), which
loops over dispatchBB().

This will allow us to simplify the interpreter and make it work one frame at
a time, i.e. an attempt to call a function would return control back to the
unique stub that sets up the frame and reenter handleResume() (or jitted
translation). This will allow us to eventually combine native and VM stacks.

Reviewed By: ricklavoie

Differential Revision: D69624622

fbshipit-source-id: 53db9dd113f9753fe08916376cd4b31f8270fc82