Use named creator for Thrift Flag observer
Summary:
Another problem to solve is the Thrift Flag Observer. It internally uses multiple layers of observer:

 {F1975238348}

But are we really interested to know these internal observer dependencies in the Thrift Flag? Probably not. Instead, for our interest we probably just need a `Thrift_Flag_{type}_name` for each observer used.

So this diff makes two changes:
1. Use a named creator when creating the observer for thrift flag, which should have little mem overhead.
2. Skip building the underlying dependency graph when encountering the thrift flag node.

Differential Revision: D69791152

fbshipit-source-id: 4a61d7a426b2db6221d67522ae8c7720291b03ae