Implement delivery callback for HQStreamTransport
Summary:
I'm modifying `HQSession::HQStreamTransport::sendWebTransportStreamData` so that delivery callbacks are conveyed to the passed-in `DeliveryCallback*` if it is non-`nullptr`.

I'm also adding some state to keep track of the preface size written to the beginning of each stream, so that we can subtract it off when we fire the delivery callback.

Reviewed By: hanidamlaj

Differential Revision: D69619154

fbshipit-source-id: e9f4f4444ea3fe555fd3af6be9d47d3a779b0f81