hh_distc FFI: return streamed errors rather than final set
Summary:
When hh_server calls into distc, it periodically polls errors from socket, then writes them to the errors file to stream them. These polled errors are then dropped. Then when distc has finished running, it joins the distc thread and obtains the full list of errors.

This was not taken into account when applying warning filtering to errors: we only filter warnings from streamed/polled errors, but not from the final set. This causes improper logging of error counts, plus spurious warnings if a distc typecheck is interrupted (for example by a rebase) then resumed.

This diff makes it so that we return errors from the accumulated streamed errors rather than from the joined errors. An alternative fix would be to reapply warning filtering to the final set, which we want to avoid. The next diff will add a probabilistic check for error consistency. The probabilistic check is implemented with a sampled JustKnob, set to 1/10 (quite high given how rarely distc is triggered).

Differential Revision: D69985879

fbshipit-source-id: 2674c0523d28fe0ddb272c73fb03a9de187fd228