Hover shows uninstantiated type
Summary:
See test plan for a before/after picture

Basically instead of showing the instantiated type of methods, functions and properties, we show the signature as declared, then show the type substitution.

To do so, no big change in the tast or the typechecker. Instead, we just reconstruct the instantiation from the decl type of the method/function/etc. and the type from the TAST. This is done in Derive_type_instantiation.

This is all gated behind a flag to be able to observe performance changes.

Collateral impact: any `as mixed` shown anywhere will be deleted.

Reviewed By: mheiber

Differential Revision: D69662474

fbshipit-source-id: ae84843d88bec46fe535d69fe1b64d5fc8585656