Updated the build.sh to get Proxygen installed correctly on CentOS dev severs
Summary:
Current build.sh script of Proxygen fails to build on the CentOS dev servers (which are also used for the Confidential virtual machines). On investigation, I found 3 errors:
a). lsb_release is not installed by default
b). Reqiures installation of fast_float-devel
c). Requires setting up of proxy server settings so that it could seamlessly download the github repos.

The diff does not change anything for the existing systems, but instead make it work seamlessly for the CentOS servers with proxy server.

Reviewed By: kvtsoy

Differential Revision: D69965614

fbshipit-source-id: 1eebb1b8b3accf6aa7f130a805539cd516e7a864