Introduce __NoDisjointUnion attribute and check
Summary:
This diff introduces a new attribute. When the attribute, `__NoDisjointUnion` is attached to a class/function/method type parameter, calls the the relevant entity are warned against if the type inferred for them is a union type and the types in the union are disjoint.

This is useful for marking common mistakes in comparisons, e.g., `===`.

This feature is gated behind `no_disjoint_union` unstable feature. In the parser, we mark it as `OngoingRelease` because there are no runtime impacts, so it'll be sufficient to control the release through the `.hhconfig` file.

Reviewed By: andrewjkennedy

Differential Revision: D69859007

fbshipit-source-id: 054f61c7f21f1c099b355c1f769c201ced0cf9c5