Record the et free variables at the splice that contains them
Summary:
If a splice contains an expression tree with a free variable, we will want to be able to pull that variable in from the enclosing ET.

Switch from tracking a HashSet of variables to a HashMap where the keys are the variables, and the values are the position they occur in. This will let us give reasonable errors when there's a free variable in a nested tree where the enclosing ETs don't bind it.

Reviewed By: mheiber

Differential Revision: D69535115

fbshipit-source-id: 92d84128a06eb56deadc3296fcc6fb17b026768c