IoUringBackend: issue sync fixed len recvzc requests
Summary:
Add a way to synchronously issue a fixed length recvzc read from IoUringBackend (as opposed to a long-lived recvzc read w/o a limit via AsyncIoUringSocket).

Use the classic ReadCallback API and copy the data. This API is intended to be used for peeking only so the copying is fine. If it is a problem in the future then I can change it to return an IOBuf.

# Context
This and the rest of the stack is needed to support peeking done at the crypto layer in wangle::Acceptor to determine things like StopTLS and at the Thrift layer to determine the protocol. 9 bytes are peeked for crypto and 13 bytes are peeked for Thrift.

The peeking happens early on before an AsyncIoUringSocket is created on top of AsyncSSLSocket. Prior to this, the only way to read a socket configured for io_uring zc rx is via AsyncIoUringSocket.

Reviewed By: dmm-fb

Differential Revision: D69964064

fbshipit-source-id: 2df08d6063bcba558bb8d124365797920137aaa4