sandbox jumpstart relative paths shouldn't have leading slashes
Summary:
When the root does not have a trailing slash, as php_root() creates,
then splitting the absolute path after the root results in a relative path that
has a leading slash. We later read these relative paths from the profile and
join with root and a "/", eg `makeStaticString(root + '/' + relpath)`.

To support this, just ensure that when we create a relative path we support
roots with and without a trailing slash.

Differential Revision: D70198439

fbshipit-source-id: 6c539b3111675bcb70f6b31bdc34eb995f91d618