Adjust test "do_not_expand_for_intersect.php" to avoid GADT issue
Summary:
This test is trying to make sure that when we intersect with a case type with where clauses we don't expand to the variants (because the variants are conditional, it's not safe to do so unconditionally).

However, this specific example runs into an issue with GADT-style reasoning that I'll add in D64611956. Essentially if `$x is int` we know `MyInt < T` and so we know that `int < LiftableTo<T>` which allows `int & LiftableTo<T>` to simplify to `int` just by the "is one a subtype of the other" rule. (The class-ish GADT logic is susceptible to the same kind of issue.)

This diff adjusts the test to test what I want while avoiding this simplification.

Reviewed By: dlreeves

Differential Revision: D69672616

fbshipit-source-id: aef512b94dcfbb712bf0b4736ac6cec2050b39b2