Fix move assignment for TypeErasedValue
Summary:
The current implementation of move assignment is completely broken if the two `TypeErasedValue` instances are of different types. the `move_assign` in the vtable is called with the wrong type!

This diff implements move assign in terms of destroy + move (similar to move constructor) and removes `move_assign` from the vtable (which already has very limited usefulness)

Reviewed By: evanjzou

Differential Revision: D70298897

fbshipit-source-id: 39f2001d1f3338b35068c437c478360d7dbcb7aa