Thread 'line_agnostic_hash' through to diagnostics
Summary:
Adds the `line_agnostic_hash` of a `User_error.t` to the LSP diagnostic message in the `data` section.

There is some unfortunate plumbling here because the hash is generated on the Rust side of the FFI and we have no way of passing a  'finalized' `User_error.t` over the boundary because the Rust implementation of `Pos.t` is not parametric in the file represetnation. As a consequence, we have to update the `ClientIdeMessage.diagnostic` type to store the hash before we 'finalize' the error .

Reviewed By: mheiber

Differential Revision: D70319210

fbshipit-source-id: 811f611cdd2a17b49d289eac055e8153402cef43