Avoid emitting a COPY relocation to _r_debug
Summary:
Accesses to ```_r_debug``` in translation units compiled without PIC/PIE in the small
code model force a COPY relocation to be emitted for that symbol. This results in it
being "defined" twice -- once in ld.so, and another in the application -- at runtime.
LLDB uses that symbol to find the list of loaded modules, and it always uses the one
in ld.so.

This diff moves the access to ```_r_debug``` to a .cpp file, which is then compiled
with "-fpic". This means the access will require an indirection (which shouldn't be
an issue as ```get_r_debug()``` shouldn't be hot), but LLDB will be able to find all
loaded libraries.

Reviewed By: Gownta

Differential Revision: D70347301

fbshipit-source-id: b75c1aed27ef55e40b7740f5892352f101766ce8