Prevent gmp_pow from allocating too much memory
Summary:
By passing a huge exponent to gmp_pow, you can cause the runtime to
allocate an arbitrary amount of memory. Even worse, since GMP
internals don't allocate from the request heap, it won't trigger an
OOM.

Unfortunately there's no good solution for this. Ideally we would hook
up GMP internals to the request heap and trigger an OOM if it tries to
allocate too much. Unfortunately the GMP memory API does not provide
for allocators which can fail. So there's no way to raise an OOM.

The best we can do is cap the size of the exponent you can pass to
gmp_pow. This limit is pretty arbitrary, but 2^32 seems reasonable.

Reviewed By: ottoni, OctavianGuzu

Differential Revision: D69748417

fbshipit-source-id: 4211643dc99734869e98ac65be04289f5bd848be