Fix ASAN builds
Summary:
ASAN builds were failing because our TC mmap started to
overlap with the "shadow memory" range that ASAN allocates. This might
have been going on for some time, but it wasn't noticeable because we
used MAP_FIXED. MAP_FIXED doesn't report an error if some part of the
request tramples on a previously allocated range. As a result, in ASAN
builds we might have been overwriting ASAN state with code (and
vice-versa). This only became a hard failure when the overlap became
larger enough that a write in the TC caused ASAN to try to write past
it's shadow range and into an unmapped region (which then segfaulted).

Fix this two ways. First, use MAP_FIXED_NOREPLACE instead of
MAP_FIXED, which does error if any part of the requested range
overlaps an existing range. In normal builds, this is a hard error (it
shouldn't happen at all). In ASAN builds, we use that failure to infer
that we're overlapping with ASAN state and instead halve the TC size
and try again. This isn't a big deal because when using ASAN, we don't
expect to run the JIT that much (or at all).

Reviewed By: jano

Differential Revision: D70468701

fbshipit-source-id: 4384229d665de6126068e518c018c4d003379141