Explicitly handle class <: classname instead of rewriting
Summary:
The projection reasons ended up not being useful for tracing flows of class to string exposure, so this change removes the rewrites and directly implements the transitive subtypes. This saves on constructing a classname locl type for rewrite. It also skips decl lookups for variance checks on C::class values flowing to `classname` (when the `class_class_type` flag is on).

For ease of review, I'm writing out the (unchanged) declarative rules:
```
1) T <: U => class<T> <: class<U>

2) {} => class<T> <: nonnull

3) T <D: dynamic => class<T> <D: dynamic

// when flag enabled
4) T <: U => class<T> <: classname<U>
```
the corollaries of rule 4 are
```
1) T <: U => class<T> <: typename<U> // by classname's bound
2) class<T> <: string // by typename's bound
3) class<T> <: arraykey // by (2)
4) class<T> <: Stringish // ^
5) class<T> <: XHPChild // ^
6) {} => class<T> <D: dynamic // by (2), string <D: dynamic
```

Reviewed By: viratyosin

Differential Revision: D70531161

fbshipit-source-id: 307a3e7a509b5518cb5998573bd75946b483829a