Support desugaring for expression_tree splices that act as macros
Summary:
Update desugaring of splices which contain nested expression trees with free variables. We want to treat these as "macros" by wrapping them in lambdas to delay their evaluation. We also call a `macro_splice`, or `async_macro_splice` visitor (instead of a `splice` visitor) that gets the lambda and list of free variables. This visitor can then invoke the function once the expression trees for the free variables are ready (according to its own visitor setup).

Update to the type checker to come in subsequent diff.

Reviewed By: mheiber

Differential Revision: D70390448

fbshipit-source-id: ea3cc9d9f4022085724db8b9263ab5bbb33edf54