Remove Skip Code Annotation
Summary:
**Stack**
The pz2 code config is an outliner in that it is the only pz config that is not part of the asset config policy system heirarchy (https://fburl.com/code/jrxvxoct).  This prevents us in uniting the privacylib pz1 and pz2 configs, like what was done for entschema.  This stack replaces the existing pz2 code config with a new pz2 code config that is within the pz hierarchy.

**Diff**
`PZ2CodeAnnotationSkip` is no longer used as it was replaced with `PrivacyLibPZSettingSkip`, so it can be removed.  This now makes `PZ2CodeAnnotationStatic` the only subclass of `PZ2CodeAnnotation`, so we can just merge them into one class.

Differential Revision: D70479014

fbshipit-source-id: 3495b5b52f984fe7cf18c13884746d04e044dcc4