Add late override support to Thrift pluggable functions
Summary:
In some cases, such as Python debug builds, we may not be able to guarantee that pluggable functions are properly registered before they are used. This is also not necessarily a serious error that warrants terminating the program.

This change introduces opt-in support for allowing late overrides of Thrift pluggable functions, which is made available using the new `THRIFT_PLUGGABLE_FUNC_REGISTER_ALLOW_LATE_OVERRIDE` macro.

This doesn't add any functionality to support intentional late overrides as I don't see any great need for that at this point (expect in the new unit test, which is pretty ugly).

This implementation does add some overhead due to introducing a new member variable to `PluggableFunction`. I considered using a template approach to avoid this, but that would have required allowing late overrides to be specified in the declaration and specifiying it in the registration made more sense to me. Happy to rework this though.

Reviewed By: andriigrynenko

Differential Revision: D70486907

fbshipit-source-id: 29948a536c839bd93111a120e97a75d3a1dc69b5