Use in-memory depgraph for invalidation of decls in IDE daemon
Summary:
It turns out that even without loading the saved state depgraph, the IDE daemon does have a depgraph at hand which it can use to invalidate only the necessary folded decls.

This diff:
1. adds the necessary 'extends' dep edges for proper decl invalidation
2. queries the depgraph to decide what to invalidate.
3. adds logic to maintain a dep resolution table mapping dep hashes to class names.
4. flushes the dep edges at a strategic time (just before querying the depgraph)

For __1.__, it turns out typechecking alone does not add all the dep edges we need for decl invalidation. We do need to add more edges. More specifically, for each class decl we make and store, we need to record the dependencies of that class to its parents. That way, whenever a class changes, we can traverse those dependencies recursively to collect all the descendants of that class so we can invalidate them.

__2.__ is pretty straightforwardly done in helper `invalidate_folded_decls_flush_deps`

For __3.__, the resolution table is a hash table added to the local memory under field `dep_table`. Entries are added just before we add the `extends` edges in 1.

This diff is behind a flag.

Reviewed By: mheiber

Differential Revision: D70002932

fbshipit-source-id: 15549a829fbb4611d0ca9b0772b0dc52e753edff