fix nullptr-dereference in SocketAddress
Summary: When `getaddrinfo` fails, `SocketAddress` attempts to describe the error and throw an exception with that description. It includes the content of the `node` or `host` param, which is a C-string, without checking whether the param is nullptr. When `SocketAddress::getAddrInfo` is called with a nullptr param but where its internal call to `getaddrinfo` fails, this triggers a segv crash.

Reviewed By: knekritz

Differential Revision: D70706294

fbshipit-source-id: 0f4144bcb670c4e8e24dd014c9bc0c4ccad06981