MapPatch: use `std::forward_list` on incomplete types instead
Summary:
We are trying to migrate fbcode to use libc++. `std::vector<T>` requires T to be complete before any member of the class template is invoked. Here we are instantiating `MapPatch` and `~vector` is used.

We found a few standard library types including `list` and `forward_list` that don't have this problem. `std::forward_list` is chosen because of its minimal theoretical overhead.

Reviewed By: thedavekwon

Differential Revision: D70726443

fbshipit-source-id: 970caaa05265db3d5edbfadb0a8c064c20790b1e