Prelimaries for type checking nested expression trees with bindings
Summary:
When a macro splice is introduced, it introduces fresh type variables for the free variables in nested expression trees. The splice is also hoisted out of its context and assigned to a temporary. We need to stash the mapping of nested bindings to their types to check for consistency the enclosing expression tree environment where the splice is used.

For example,
```
 ExampleDsl`{$x = 1; return ${ExampleDsl`1 + $x`};}`
```
turns into (simplified)
```
$splice0 = ${() ==> ... ExampleDsl::intType()->__plus($x) ...};
...
  $x = ExampleDsl::intType();
  return ${$splice0()};
...
```
At the point where `$splice0` is used, we need to see that `$x` is bound to a client `intType`, and relate that to the type of `$x` as used in the nested ET. To do this, we add a field to the `local` record that stores information about each local variable, and store the information about the free variables ($x) of the splice in the binding of the temporary, so that we can retrieve it when checking the use of the splice.

This strategy relies on the type checker being able to detect what the desugaring is doing, so there is some coupling. The alternatives is to embed the information in the type of the splice, either by extending the `Spliceable` type or adding a new type that wraps the spliceable type. The first is visible to the programmer, which we want to avoid, and the second complicates the whole type system and is also not robust because of needing to look through the wrapper where ever types are inspected.

Reviewed By: dlreeves

Differential Revision: D70566100

fbshipit-source-id: d1f77e2ac786f8d6c05c3bd93a50c545ac2b2166