Fanout tests: check expected included in actual for series of changes
Summary:
For fanout tests, we have a "series of changes" mode where the test file specifies a series of repo changes. In that case, hh_single_fanout outputs the fanout as
```
<fanout of change1>

<fanout of change2>

...

<fanout of changeN>
```
Then verify.py just checks that the content of .exp is equal to the output.

However, we'd want to specify the expected fanout as the minimum fanout, then our fanout algo is free to be more pessimistic and include more stuff.

With this diff, the new output looks like
```
<fanout of change1>
--
<fanout of change2>
--
...
--
<fanout of changeN>
--
```
Then verify.py splits this output at `--`, then for each section, it checks that the expected fanout is *included* in the actual fanout.

This is in preparation of the next diff, which will use a more pessimistic fanout algo.

Differential Revision: D70706326

fbshipit-source-id: 012871739ae70fded96baa6225808db9612d2547