Use atomic for active_fetch_action_
Summary:
In D70333626 I attempted to not make active_fetch_action_ an atomic.  This variable is only changed in the mysql event base and only needs to be read (very occasionally) outside of this thread.  My thought was that inside the thread there would be no protection and from outside the thread we would schedule a function to read it inside the main thread.

Unfortunately this broke https://www.internalfb.com/intern/test/844425027465877?ref_report_id=0 as the test causes the variable to be read outside of the event base thread, but the event base thread is in a blocked state, so the code never runs and the test just freezes.

The correct solution is just to make the variable an atomic.

Differential Revision: D70713529

fbshipit-source-id: bd61baca28d337fa99cb8d389619501937af4ed2