Determine the builtin ABI based only on the C++ signature
Summary:
When passing arguments to a C++ builtin, the only thing that determines the ABI
(how the input should be encoded) is the actual C++ signature.

The previous logic was quite bizarre, it used only partial information about
the C++ signature and then filled the gaps based on the Hack signature, whether
the Hack type was representable by a unique DataType, or it was one of the
DataTypes from a random list.

The C++ and Hack signatures still obviously need to be compatible, but that's
what the verifier is enforcing.

While there, make it less fragile to pass floating point numbers to builtins.
Although it seems to be always the case due to entry assertions, do not rely
on the type to be inferred as TDbl.

Differential Revision: D70861196

fbshipit-source-id: 2f19d4182d01d9da24fff0f94ce93f77cb759311