Fix bug in over_x_ms outliers
Summary:
So for this counter we have a vector of `thresholdMs`, let's use {100} for this example.

In the constructor we create a outlier timeseries Ms map where we multiple the threshold by 1000
https://www.internalfb.com/code/fbsource/[47188bc4d17b0b937a53e4b70ce5d96ff09532f1]/fbcode/hphp/runtime/base/profiling-counters.cpp?lines=46

Then when this is logged, we multiply the entry by another 1000
https://www.internalfb.com/code/fbsource/[47188bc4d17b0b937a53e4b70ce5d96ff09532f1]/fbcode/hphp/runtime/base/profiling-counters.cpp?lines=57

and compare against the value in us.

So 100ms becomes 100 * 1000 * 1000.  we're out by 1000.

This code changes `mOutlierTimeseriesMsMap` to actually be in `ms` not `us`.

Reviewed By: mdko

Differential Revision: D70907445

fbshipit-source-id: 25e8536ab0dd4d442138bf3f6db75bb4cda18e08