Extend __NoDisjointUnion to cover nullable of disjoint unions
Summary:
When the typechecker infers certain types, they get normalised, particularly unions and nullables.

A consequence of that is if we infer the type `?string | int` or `?string | ?int`, they both will get normalised to `?(string | int)`. This normalisation is correct.

This normalisation challenges the `__NoDisjointUnion` annotation that heuristically detects the cases where a type parameter gets bound by multiple pieces of data with types that are disjoint. This typically matter in case of comparisons.

While `?string` is not disjoint from `?int`, it is disjoint from `int`. After the normalisation, we cannot distinguish the two cases (and without the normalisation, we'd be distinguishing it heuristically in any case!).

This diff makes the decision that when we see `?(ty | ty')` or similar, and if `ty` is disjoint from `ty'`, we flag it as a disjoint union even if in reality `null` might be at the intersection of the two sets.

The reasoning for this design choice is that even if such an intersection exists for an API that is marked with `__NoDisjointUnion`, the chances are that they made a mistake and didn't want to compare `?int` to `?string`. So even in the case of false positive, we detect a bug.

Reviewed By: viratyosin

Differential Revision: D70784852

fbshipit-source-id: 6a49c54dbb9ee9e7ee130b1ee9445455ed254b5a