Don't require ids to be in the experiment
Summary: We currently decide to log the experiment ids only when they already exist in the context. This may not be the case right before we run the genExecute method. We need to log cases where we have no ids before calling genExecute, but we have ids in the context after.

Reviewed By: charnahalau

Differential Revision: D70930092

fbshipit-source-id: 3556975012d5f201ee84ea780d8ce72a104f1f7d