Allow native_object::lookup_property to throw
Summary:
For migration reasons, we need to expose properties which invoke functions underneath the hood. These functions can throw errors.

For example, `{{enum:name}}` should become `{{enum.name}}` and the lookup for `name` should trigger a function call. There are some existing mstch methods that throw that we need to account for that can throw.

Reviewed By: vitaut

Differential Revision: D70757043

fbshipit-source-id: 09137ba2fad12e3d14fc9dca8ce1d3ae3e8453a0