Return found object + parent in eval_context lookup
Summary:
In the next diff, I've added a `self` argument to function calls. That is, the expression `(a.b.c ...)` where `c` is a `native_function` will receive `b` as the `self` argument.

This is similar to how JavaScript's `this` works. It allows us to emulate member functions without additional machinery.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/this

> The value of `this` in JavaScript depends on how a function is invoked (runtime binding), not how it is defined

Reviewed By: vitaut

Differential Revision: D70763705

fbshipit-source-id: e869f17760570aff219784fb9cb9df196564180e