Fix bug in validating enforceable hint for classes with reified generics
Summary: When we encounter a reified tparam, we'll set `inside_reified_class_generic_position` in the accumulator state and recurse to visit the type, however, when we finish, we continue to use that same updated accumulator state. However, this updated value of `inside_reified_class_generic_position` is not necessarily correct moving forward. This error was resulting in both false positives and false negatives for checking certain class hints with multiple generics.

Reviewed By: mheiber

Differential Revision: D70927718

fbshipit-source-id: 40cf13000224fca27e310d7d3b1082a5329f220c