Add possibility to write whole Cursor serializer into struct field
Summary:
Sometimes writing fields in order is not really possible because tha earlier field values depend on the later field values. And for some cases changing the ordering of the fields is not easy and may lead to long and painful migrations.

This change basically adds an API into the StructuredCursorWriter to be able to accept a CursorSerializationWrapper that is prepared separatelly and would contain the content a whole stucture that is already serialized at that point. This will result in the serialized IOBuf being chained to the existing IOBufQueue.

Reviewed By: iahs

Differential Revision: D71001961

fbshipit-source-id: db002a41bdd4edcbc5adc6dd9ab078488c64e08c