Leverage heterogeneous access for `F14HashedKey`
Summary:
`F14` containers support heterogeneous access, we can use it to avoid
duplication for `F14HashedKey`.

Simplification doesn't come without downside: heterogeneous access should be
enabled to use `F14HashedKey`, before it wasn't the case. In reality, we use
`F14HashedKey` almost exclusively for with string-like keys, where
heterogeneous access is on out of the box, so this shouldn't be a big problem.

Reviewed By: Gownta

Differential Revision: D70641009

fbshipit-source-id: 24ed425f197445adc98f73895111b9f15dccf4cb