Fix lambda subtyping with polymorphic function as expected type
Summary: I'd introduced a bug by instantiating polymorphic function types when they appear as the expected type under bidirectional typing rather than freshening and binding. Since we already do this work in subtyping, this diff rolls that back and types lambdas without an expected type when that type is a polymorphic function.

Reviewed By: mheiber

Differential Revision: D70968631

fbshipit-source-id: c5d34a07eff50b2e4ebf0423f8171a3fd9b3b421