Fix Zlib, Gzip & Bzip2 truncation bug
Summary:
The `ZLIB`, `GZIP`, and `BZIP2` codecs truncated input sizes larger than 4GB to 32-bits. Fix the bug.

* Add a `chunkedStream()` helper that calls the streaming implementation in a loop with input / output size bounded to the `chunkSize`.
* Wrap the Zlib & Bzip2 streaming implementations in `chunkedStream()` to guarantee they never see sizes > 4GB.
* Add explicit casts to `uint32_t` with `DCHECK()` that we haven't truncated to the Zlib & Bzip2 streaming implementations.

The logic to implement `chunkedStream()` correctly is a bit complex, so please review carefully. Implementing this logic in each of the implementation functions would be slightly simpler in each case, but IMO the sum of the logic would be more complex.

NOTE: I've limited the chunk size to 4MB rather than 4GB so that the chunking code gets executed regularly. This bug has been latent for nearly 7 years, which shows that sizes > 4GB are rare.

This bug was originally fixed in D1702925, and then re-introduced in D5169338 (zlib) and D9484325 (bzip2).

Reviewed By: felixhandte

Differential Revision: D70936552

fbshipit-source-id: 72ec11148613d6ed9643da47946bf4bf89a6faa8