Remove Read/Write String APIs from (De)Serializer
Summary:
`string` and `[]byte` are very easy to convert to/from in Go.

No need to maintain additional API for `string` type.

If we look at `encoding/json` package - only `[]byte` API is provided:
https://pkg.go.dev/encoding/json#Unmarshaler

That's the idiomatic way - only having `[]byte` API for implementations that deal with encodings.

Plus, `string` API only has a handful of callsites. It just needs to go.

Reviewed By: leoleovich

Differential Revision: D70994659

fbshipit-source-id: 6b81aacf8664ade58fbfb482ee891ab3150e9566