Add support for DSL::shapePut to write shape fields
Summary:
This diff provides the `<DSL>::shapePut` API inside Expression Tree backticks, which is desugars to `Shapes::put` (and is semantically equivalent).

Special handling is needed for this function in the virtualization, because the DSL shape type has to be converted into the Hack shape type so that the type checker can reason about it and extract the field type.

This change only affects Expression Tree virtualization (type checking), and not desugaring. The function is desugared as a normal static function call and it is up to the DSL implementer to provide an implementation for `<DSL>::shapePut`.

```
DSL`DSL::shapePut(shape('x' => 1), 'y', 'test')`
// inferred to DSLShapeType<shape('x' => DSLIntType, 'y' =>DSLStringType)>
```

Differential Revision: D69304418

fbshipit-source-id: 63e2d3b28c8a2a7dfe87f98f4b593819b83930fb