Implement a Mini Interpreter
Summary:
This implements a simple interpreter for a subset of Hack that is suitable for running within hh_client. For the parts of the language it supports it will produce a fixed set of values. If it ever encounters a language construct it doesn't know, an error value will be reproduced instead.

Though not implemented in this diff, the idea would be to have the type checker validate that code that would be interpreted fits within the desired subset. We will leave that as an exercise for future diffs.

Reviewed By: enetsee

Differential Revision: D70025535

fbshipit-source-id: 9c67263cd0dddf1557b2cd181cc2d7fa53ad9cb6