Allow ClassPtrs to go to Strings
Summary: The interpreter resolves strings and classnames to different values. This is because classnames are special and are strings intended to refer to defined classes. Though within Hack it is expected to use classnames as string as well. For the interpreter we introduce a `Value.str_ish` to indicate places where we expect something that can act like a string. For the case of `Value.ClassPtr` we will covert it to a string which is the name of the class

Reviewed By: enetsee

Differential Revision: D70805272

fbshipit-source-id: 90ebe07093c9c06a77dc9e9704438c5050738b0f