Move destructor to the end of the file to make types complete when instantiating vector
Summary:
We are trying to port thrift to build on libc++.

libc++'s vector is stricter such that it requires the type argument to be a complete type when its destructor is referenced. This can usually be solved by outlining the destructor to later in the file, which is what we have done here.

Reviewed By: iahs

Differential Revision: D71133898

fbshipit-source-id: 2f2c9fc41279e17991224c942bdfe348e3940d00