profiling-counters: add AVG and COUNT stats types
Summary:
While updating web-compare I realised that we can't calculate the average `hhvm.memcache.get_duration_us` as we don't have the count. Because these are not 1:1 with requests we can't use that count.

We can't just add the export type `::COUNT` because this is called once per request and gives a far too low a value.

This diff pulls the count from the `ServerNote` and uses it to add and aggregated value to the timeseries.

Reviewed By: Flameeyes

Differential Revision: D71055563

fbshipit-source-id: 63e2500b4ac0b76ad0df195b9963d9b9a166944d