Refactor to prepare for inline IRUnit stitching
Summary:
Currently, we generate the callee `IRUnit` when calculating the cost of inlining for the first time (which then gets cached). This diff refactors to return back to `irGenTryInlineFCall` the callee's `IRUnit` whenever `Cfg::HHIR::EnableInlingPass` is true or we need to calculate cost for the first time.

In order to support that, I had to move the generating `IRUnit` out of `computeTranslationCostSlow` into a new function `computeInlineIRUnit` which returns whether to inline the callee region and also returns back the unit.

Reviewed By: paulbiss

Differential Revision: D70131260

fbshipit-source-id: 41b3e30d5f4100cb8ed4e88ea322ea58e738082a