Add serializer V2 implementation
Summary:
A more idiomatic and faster-performing API for serialization.

# Idiomatic:
We are following the same pattern as other Encoder-style APIs in Go lib - e.g.:
https://pkg.go.dev/encoding/gob#Encoder
https://pkg.go.dev/encoding/json#Encoder
This means - accepting a `io.Writer` in the constructor and providing an `Encode` method.

# Faster-performance:
As shown by benchmarks downstream - this updated API performs faster due to fewer allocations/writes.

Reviewed By: podtserkovskiy

Differential Revision: D71160874

fbshipit-source-id: 3dceef4882751e5187c3d5c174b3c92568daea91