Box all large fields in Protocol Value
Summary:
This reduces the static layout of Protocol Value from 72 bytes to 24 bytes, we can further reduce this by removing the inheritance of Protocol Value. This is crucial change for the container usages, as there is a significant overhead of memory usage compared to container of primitive. However, this trades-off with extra indirection when accessing these fields.

For example, if you use `vector<i32>` the allocated size will be `4 bytes * capacity`, but if you `vector<Value>` to represent `vector<i32>` the allocate size will be `72 bytes * capacity`, which is 18 times larger allocation. This diff offsets with boxing large members by reducing the static layout size of Protocol Value.

It results in a considerable performance increase when encoding/decoding `Protocol.Object`, at the cost of small regression when encoding heavily nested struct data. Note that is partially alleviated by the next diff in the stack (Removing class inheritance from the ObjectWrapper/ValueWrapper).

#buildall

Reviewed By: iahs, thedavekwon

Differential Revision: D71132089

fbshipit-source-id: 7fdd622854649034f0059a21f52cc87af8305408