Support memory_resource to allocate IOBuf
Summary:
Some applications require tight control of all the allocations backing `IOBuf` chains; currently it is possible to externally allocate the user buffer and wrap it with `takeOwnership()`, but the storage for the `IOBuf` node is always allocated using the system allocator.

This diff introduces initial support for providing a `std::pmr::memory_resource` to `takeOwnership()` and `clone()`, thus allowing to control all the allocations in the common lifetime of a `IOBuf`.

This is going to be a niche feature, so in order to avoid increasing the `IOBuf` footprint or adding other overheads when the feature is not used, the pointer to the `memory_resource` is stored as a conditional prefix of `HeapStorage`.

The `memory_resource` is intentionally not propagated on `clone()`: this allows to relax the thread-safety requirements, since allocations are only done in the methods where the `memory_resource` is explicitly provided.

Reviewed By: Gownta

Differential Revision: D69181515

fbshipit-source-id: 630b079b07570d1b031f5b139bc5c1555655fa73