Add support for soft __DynamicallyReferenced classes
Summary:
For migration, make it so <<__DynamicallyReferenced>> can take a single int parameter. If this int is passed, we say it's soft dynamically referenced and log notices at the provided sample rate.

So, `__DynamicallyReferenced(0)` will have the same behavior as `__DynamicallyReferenced` except it will go down the slower path of checking the user attribute at runtime instead of not emitting the IR op.

Differential Revision: D71154722

fbshipit-source-id: ccf50e4fe8dbebf93d9f6c5ad386459456883e97