`SafeTask`: like `coro::Task` but restricting unsafe aliasing
Summary:
Explained in the file docblock.

This is useful independently of `async_closure()` insofar as you can default to writing `ValueTask` everwhere, and this will catch a large fraction of (real or potential) dangling reference bugs at compile-time.

There are two reasons to introduce the much heavier machinery of `async_closure*()` later in this stack:
 - It provides a robust solution to the "async cleanup" problem. The `async_arg_cleanup` generalization  allows the closure to await multiple cleanup steps on exit (like `co_scope_exit`), but the cleanup is actually enforced to be memory-safe (unlike `co_scope_exit`).
 - It supports compile-time-checked "safe" reference-passing from ancestors to descendants. From a user's perspective, they just pass `async_arg`s `by_ref` into child closures, and it works. Or when it fails to compile, this means there may be a memory-safety bug.

Reviewed By: ispeters

Differential Revision: D61995850

fbshipit-source-id: e7e1846285fb89d32f2baa34bd6c3d9dfb91ff73