Regen www with codegen change in D70428261
Summary:
as title

There are some tests that mock `genAwaitResponse` or validate that it was invoked. Since the end result is that `genAwaitResponse` will return (resp,header), I decided to move the mock to a helper class to hide that detail from the user. For now, mock uses `genAwaitResponseWithReadHeaders` until we move everything back to `genAwaitResponse`

 allow-large-files

Reviewed By: yoney

Differential Revision: D71223284

fbshipit-source-id: dec17bb9568ead4b97e06a5d24aab6b6e4b22645