Fix nested expression tree macro splices with await
Summary: If we have a nested expression tree in a macro splice that uses await, then we lift the splice out and wrap it with an async function. The runtime expression just passes that function into the visitor, whereas the virtual expression needs to call it, to get the return type. We also need to await the result. However, the virtual expression itself if wrapped in a synchronous lambda, and so we can't use await. Instead we directly remove the Awaitable wrapper in type checking. This have the same effect since the virtualized expression is never executed.

Reviewed By: mweststrate

Differential Revision: D71318190

fbshipit-source-id: ea7c0f555d21b2d5f4516326896d987996e86cb4