Add support for enum values to the whisker generator
Summary:
Add support for enum values to the whisker generator. Remove `enum:size` which is only used in two places, replacing it with direct usage of the AST via whisker. Apart from simplifying code, this makes the semantics more obvious by making it clear that we are looking at the number of enum values and not, say, the range of values.

This is also a good example of how whisker will allow us to get rid of the intermediate mstch layer.

Reviewed By: praihan

Differential Revision: D71319536

fbshipit-source-id: df183621891ad5b3c6ea14005b28a5169ce76b86