Fix Arena allocator aligned size calculation.
Summary:
D62702993 added support for >8 alignment, but its size calculation was wrong. It aligned the memory pointer after allocating the 8-byte `Block` header, but did not align the earlier size calculation for the `Block`.

This could cause heap corruption in the case where the arena minBlockSize is already aligned, because the allocated buffer would be 8 bytes short. In particular, it broke in the `PosePool` after I changed the `PoseState` structure to make it 8 bytes smaller. Constructing that structure would write past the buffer and trigger a heap corruption exception in debug builds.

Reviewed By: yfeldblum

Differential Revision: D71167717

fbshipit-source-id: c7c4af9aeff0d79d772167ef2caaf6342e57e97c