Fix incorrect variable name
Summary:
It actually means whether definition has uri. It has nothing to do with reflection.

A more accurate name would be `hasUri`, though this describes what is this variable. It's better to describe how this argument changes the function's behavior.

The current naming is caused by the following legacy reason:

* Previously we only enable always-on reflection when the struct has uri (since previous direction was enabling uri everywhere), thus `enabledReflection` and `hasUri` are equivalent.
* However, we changed the direction to not enable uri, thus we rolled out always-on reflection with structs that has no uri.

Reviewed By: aristidisp

Differential Revision: D71431184

fbshipit-source-id: 163384de338e62e69bf961c7741045b00069085d