Rename ClassGetTS to ClassGetTSWithGenerics
Summary:
`TReify::foo()` compiles to a bytecode sequence that makes use of implicit string to class conversion

```
    BaseL $0ReifiedGenerics Warn Any
    QueryM 0 CGet EI:0 Any
    BaseC 0 Warn
    QueryM 1 CGet ET:"classname" Any
    PopL _1
    .srcloc 4:3,4:15;
    PushL _1
    FCallClsMethodM <EnforceMutableReturn> 0 1 "" "" - "" "" DontLogAsDynamicCall "foo"
```

D66207622 used ClassGetTS and discarded the generics vec for the intrinsic; however, since `TReify::foo()` is a more core feature, instead, I would like to factor out the class loading into its own op. This diff just moves aside the ClassGetTS bytecode to be specific to `new`, which is the only op that needs the `generic_types` field.

# RFC on the naming

Reviewed By: jano

Differential Revision: D71359237

fbshipit-source-id: 9815d44d7cd2d45e0e1a722dbf495e7944b55cf1