Override connection's query timeout if query specifies a query timeout
Summary:
When the query timeout is set in the connection option (and a proxy is used) we set the "query_timeout" query attribute and the proxy (MyRouter or Dvaar) will use this to do backend query timeouts (https://fburl.com/code/1jhodoml) for us.

Unfortunately, we weren't overriding this if the query options set a query timeout - we were only updating the client-side timeout.  That left the proxy to have a different timeout than expected.

Add code to override the query_timeout query attribute when a query timeout is specified via query options.

Also add a test to validate this (and verify that the connection's overall query timeout doesn't get changed).

Reviewed By: fadimounir

Differential Revision: D71411501

fbshipit-source-id: e5389d03f899f5de9f410c79b3b82fdd4ca7b09e