refactor ET desugaring of special shape functions
Summary: The code used to clone the arguments to special shape functions (idx, at, put), and call the rewriter multiple times to work around Rust linearity. This was because the virtual expression for these methods needs to use the non-rewritten shape field argument, but the runtime expression should still use the rewritten argument. I've refactored to defer the call to rewriting the arguments until we know which case were are in, so that we can clone the non-rewritten argument first, if necessary. It's too late to clone it after rewriting, since rewriting consumes the argument expressions.

Reviewed By: mweststrate

Differential Revision: D71479382

fbshipit-source-id: 0fad2d4235a3004afc2bb7caffa2036647ff73ad