Add new IR instruction LdClosureArg
Summary:
LdClosureArg = LdPropAddr + LdMem. This allows us to directly place a PureLoad{AClosureArg} memory effect on LdClosureArg's src.

The other alternative to adding this instruction was to place the memory effect on LdMem's src only if the src inst was a LdPropAddr and the underlying src SSA type had the class spec of c_Closure::classof. However, the class spec information is not necessarily available while JITing code. In the worst case, we will know the type of SSA to be TObj but not its class. Therefore, this solution won't always work. Although in practice, it may work most of the times because of various other stuff in HHVM that provides class information (and it does pass our tests) but it's not a good practice to rely on that information.

Reviewed By: jano

Differential Revision: D69586162

fbshipit-source-id: 1a775f10964fc9b291e69e12a9a7c0ea653599e3