Use additional sharding in ConcurrentHashMap in mcgen-async
Summary:
We're seeing some cases where caches are waiting for jit locks and taking up computation time.

Want to experiment with different types of storage structures in mcgen-async to see if this can help.

Normal ConcurrentHashMapSIMD in the existing folly_concurrent_hash_map_simd implementation has 8 bits of sharding (256 shards). I'm setting up 9, 10 bit options to experiment with, as well as AtomicHashMap and tbb const accessor options.

As part of the jit lock contention, we're also seeing cases where the same key is being attempted to enqueue multiple times in a row, so adding an option to test if reading before attempting to write  will help, so that we don't have to take a lock in the case where we exit out early.

Reviewed By: paulbiss

Differential Revision: D63569014

fbshipit-source-id: d0c6338606bbcf201aa750a17d80ec8322ab07aa