Migrate more of C++ enum codegen to whisker
Summary:
Migrate `enum:min` and `enum:max` to whisker. Rename them to `cpp_min` and `cpp_max` respectively to reflect the fact that they produce C++-specific enumerator names and not generic min and max (otherwise they would live in the base whisker generator and not in the C++ generator).

This is a good example of the logic we **don't** want to move to templates.

Remove multiple uses of the deprecated `get_enum_values` API.

Removed unused `enum:has_zero`.

This leaves just four functions in `cpp_mstch_enum`, one of which is redundant and will be removed in a subsequent diff.

Reviewed By: sadroeck

Differential Revision: D71627824

fbshipit-source-id: 122c6e87bd4c68edfdcf71aa63f015c1911592d4