Repro deprecated error under class_class_type flag
Summary:
Minimized example of an error observed with class_class_type = true from WWW.

Observe that `TPhantom` is undefined in A where it is referenced; I suspect this is the source of the problem.

Removing the generics from `I` causes both exps not to have an error.

Removing the definition of `TPhantom` from `C` causes both exps to have the deprecated error.

Reviewed By: viratyosin

Differential Revision: D71449792

fbshipit-source-id: 199b0636c04426962e11c6c68ea1ee35ac3de91c