Do not refine values in parent frames
Summary:
This is incompatible with clearForUnprocessedPred(), which clears only the
current frame. If a block with unprocessed predecessors is initialized with
a processed predecessor that refined a value that is also used in a parent
frame, but that refinement won't be available later via an unprocessed
predecessor, and that code path is the only returning path to the previous
frame, the previous frame's state will become invalid.

It is also incompatible with inline frame stitching, since there will be only
one frame state.

Reviewed By: ricklavoie

Differential Revision: D71657251

fbshipit-source-id: d955f2bbcea36a680d28d0c3e6358bfad79aab7b