Propagate the current usage hint in ifThenElse() and cond() constructs
Summary:
If the current block is hinted with a lower usage, it is never correct to
create blocks inside if-then-else with higher usage. It will just cause
random unnecessary jumps between main/cold/frozen sections.

It is common to override these hints inside ifThenElse(), but it still makes
sense to set a reasonable default. Propagate the same hint for `done` block, as
that one gets the same traffic. Deprioritize `taken` block from `Likely` to
`Neither` (they are in the same section).

Reviewed By: ricklavoie

Differential Revision: D71665809

fbshipit-source-id: 8a53837426558304a4d777dc3a77409b3c7e2a7e