Disable stress testing eager VM reg sync in debug builds
Summary:
Stress testing eager VM reg sync is harmful. It emits a ton of junk
instructions that prevent JIT optimizations from working, hiding bugs in these
optimization passes.

There is nothing special about this that would warrant stress testing. I have
never seen this logic find any new bugs, but I have seen it hiding real bugs
many times. And it is extremely annoying to work on something and see tcprints
polluted with random garbage (or even preventing the optimization you are
working on from working in dbgo).

Also, we are not actively working on adding more eager VM reg syncing. It was a
giant CPU and code size regression when we tried it the last time.

Reviewed By: ricklavoie

Differential Revision: D71678334

fbshipit-source-id: 55ed9f958504d2c8eb6a0659bf4c9bd3cd4c13ff