Fix rvmsp() liveness when calling EndCatch related stubs
Summary:
At EndCatch, we either know that VM regs are already synced (we tracked that
the only path comes from a single BeginCatch and we did not manipulate stack),
or we need to sync it (e.g. multiple BeginCatches across various inlined frames
that all end up in an EndCatch for the first inlined call).

The second mode (requires sync) is way more common. We synced the regs, but
then we apparently lied to vasm and tell it that we don't care about rvmsp().
So vasm could have optimized it away, and even reused `$rbx` register for a
computation within the catch block, causing the stub to receive a random
garbage. Unwinder then tries to decref random stuff in memory and segfaults.

This is what happened with D51295130, but mysteriously only in live
translations.

But we should be crashing even without D51295130. Still trying to understand
why we don't. Whenever there is a BeginCatch in an inlined frame, it pops and
decrefs everything from the stack, then kills the whole frame, then continues
until reaching the top frame, where it ends at EndCatch. If we don't properly
sync rvmsp() there, the unwinder should see a ton of garbage on the stack.

Reviewed By: ricklavoie

Differential Revision: D71680187

fbshipit-source-id: 3b6a14b16784ea2a9b43281cef793caadc494a1f