Route Hack exceptions to exception handlers
Summary:
If there is an exception handler for the current SrcKey, Jmp from catch blocks
to a shared handler instead of spilling all inlined frames and terminating it
with an EndCatch. This shared handler routes Hack exceptions to the exception
handler and unwinds the frame for C++ exceptions with EndCatch.

This kills almost all remaining InlineCall instructions and reduces code size
by deduplicating spills.

It should also slightly improve performance of exception handling, but that is
negligible.

If the EH handler is not defined in the translated region, we will emit a side
exit, but it will be shared across all side exits from the same frame. Even
without sharing it is still less code than an equivalent sequence of
InlineCalls.

This could be improved in the future by adding reachable trivial EH handlers on
demand to the region being translated. Even if these edges are cold, including
them could further reduce the code size.

Reviewed By: ricklavoie

Differential Revision: D51103576

fbshipit-source-id: fc1047ffa3c449dab375964ca8b335a06a359b88