Handle exception handling of failed inlined unawaited async functions in JIT
Summary:
When an inlined unawaited async function throws an exception, route it to the
endCatch inline target. Handle it by wrapping it into a StaticWaitHandle object
and continuing at the next bytecode via a side exit.

This situation is common in concurrent blocks where returned Awaitables are not
individually immediately awaited.

That's also why we side exit instead of jmp. We similarly side exit when
suspending execution. This allows us to to more efficiently handle the common
runtime behavior of all async functions in a concurrent block succeeding
eagerly and optimizing away most of the concurrent block logic.

Kills another spillInlinedFrames. After this diff there is only one left.

Reviewed By: ricklavoie

Differential Revision: D71689167

fbshipit-source-id: a7ca157e4785335a44a5d92f69ae584bfb46f617