Merge fbcode and xplat copies of codemod
Summary:
We already merged most of the Thrift compiler between fbcode and xplat, now do the same for the codemod. Update the call sites with a codemod of the form

```
fbgs -l 'thrift/compiler/codemod:' -f '^fbsource/.*/(BUCK|TARGETS)$' | \
  xargs sed -i 's|//thrift/compiler/codemod:|fbsource//xplat/thrift/compiler/codemod:|g'
```

Remove the copy of `thrift/compiler/codemod` from fbcode since it's no longer used.

For users this change is a noop.

Reviewed By: yoney, iahs

Differential Revision: D71628829

fbshipit-source-id: 0a28ca123a7d35f0b04e4a07bf08a9a14c9c7769