Hoist refinement of frame context out of beginInlining()
Summary:
beginInlining() refines the frame context based on the callee. Let's hoist this
refinement into prepareAndCallKnown() so that the prologue logic benefits from
the same improved types.

This might not do much (especially since maxt fixed some (most? all?) of this
recently in D62225191. But it provides a guarantee that the ctx type is sound.

Also, short-circuit if it becomes a bottom (instead of emitting a ton of code
that gets DCEd later).

This moves some of the responsibility of beginInlining() outside. More is
needed to fix the last remaining spillInlinedFrames() call (side exit due to
guard failure in a func entry during inlining)

Reviewed By: ricklavoie

Differential Revision: D71702990

fbshipit-source-id: 6e998d2532d4e33840227fe0a871278888529048