Rename value:type -> value:underlying_type
Summary:
The Rust Thrift implementation is too eager to jump through typedef definitions in certain contexts, losing information about `rust.Type` annotations that aren't on the very last type.

Part of this is because the name of `{{value:type}}` sounds like what you want "most of the time", as opposed to `{{value:local_type}}`, which it is not.

This diff renames `{{value:type}}` to `{{value:underlying_type}}` inside the Rust code generator, without any behavior change.

Reviewed By: diliop

Differential Revision: D71666046

fbshipit-source-id: 50aba4ca11736c22e06230e2d90821afb4c58134