Union.FromValue
Summary:
This is pretty straightforward. The main difference from thrift-python is the field iteration order is key-sorted in thrift-python while it's declaration-order sorted in thrift-py3.

The logic from thrift-py3 is somewhat verbose due to special logic to not match python 64-bit `float` values to thrift 32-bit `float` arm is it would cause loss of precision.

Reviewed By: prakashgayasen

Differential Revision: D71592000

fbshipit-source-id: 9f7e79165e3536f092e1dc57f7ed67cbe6a87406