impl metadata
Summary:
The metadata setup in python is weird because there's a thrift file `metadata.thrift`, and then there's a library full of `ThriftXProxy` classes that wraps it. Unfortunately we have two copies of this library, one for py3 and one for python.

The strategy we can use to make this work with in-place is to use the thrift-python class with the thrift-python `gen_metadata` library. We just need to put hooks for this in `__get_metadata__` generated methods on each struct, as well as one in the thrift-py3 library `gen_metadata` function.

Reviewed By: prakashgayasen

Differential Revision: D71778463

fbshipit-source-id: 15f3ee6a6fa9aba4a51f2d209acbc1ae3d5b7803