simplify NewObjD when no reified targs are supplied
Summary:
This diff emits strictly less bytecode in sandbox mode. The
ReifiedInit implementation in the interpreter and JIT already handled ensuring
that we didn't invoke 86reifiedinit if a an object's class or parent class had
reified generics. HHBBC wasn't properly configured to support that. This diff
fixes the issue. This was reproduceable by running `tools/hhvm_wrapper
--compile -v test/slow/enum_class/enum_class00.php` and taking the changes from
emit_expression.rs but NOT the changes in hhbbc/interp.cpp

Reviewed By: paulbiss

Differential Revision: D70981171

fbshipit-source-id: 4f3a7eb91ac643356b1f7906913588712d0aadd9