ClassGetTSWithGenerics can push empty vec instead of null
Summary:
This simplifies emit_expression.rs when we have a type structure on
top of the stack. Before, we'd check if we pushed a null and if so, set the
type structure local to an empty vec before calling `ReifiedInit`. Now, we just
rely on `ClassGetTSWithGenerics` to push the right initialization value on top of the
stack.

Reviewed By: paulbiss

Differential Revision: D71144218

fbshipit-source-id: b95e97c56c68701d224fe3c50e35b1e2b4e1739d