Remove onRequestFinished() default implementation
Summary:
# Issue
If the implementation of `RequestPileInterface` is sliced, it may result in a call to `LOG(FATAL)` which crashes the server.

This potentialy caused issue reported in: https://fb.workplace.com/groups/677796806483229/posts/1735275540735345/

In general we should defer "non implemented" issues to the compiler, rather than to the crash during runtime.

# Change
Remove default implementation of  `onRequestFinished()` deferring "is implemented" check to the compiler which would assert pure virtual metod implementation.

Reviewed By: sazonovkirill

Differential Revision: D71844359

fbshipit-source-id: b955066c6ef5ea6871cf384b1539f76f0e1b9b65