Fix python patch that typedef is not handled correctly
Summary:
If there is a typedef, e.g.

```
typedef gen_patch_Foo.FooSafePatch Bar

struct Baz {
  1: Bar bar;
}
```

In this case we should not allow patching `bar` since it's patch of safe patch.

It doesn't work today since `gen_patch_Foo.FooSafePatch` is a typedef, and for typedef we did not set `is_patch` correctly in python codegen.

Reviewed By: thedavekwon

Differential Revision: D71910769

fbshipit-source-id: 897976a03a1b4c4385572d7f54379a4a128bfd4f