Move type_predicate and co into and-block with locl_ty
Summary:
I plan to put `locl_ty`s inside the `ClassTag` which will create a mutually recursive relationship between `locl_ty` and `type_predicate` so I need to move the definitions into the mutually-recursive and-definitions.

This requires removing the derivers and manually defining the printing and comparisons functions I was using since most of the derivers try to apply to all the types in the and-definitions and they aren't compatible with all the types there.

Reviewed By: dlreeves

Differential Revision: D71941113

fbshipit-source-id: a10c9ece6044845bf8f9d0d6809e4def551202f5