Change ET macro splice semantics
Summary:
Previously, macro splices were wrapped in a lambda to delay evaluation. Now we do not wrap them and execute them eagerly.

We also insert a call to `DslName::lift` around the contents of every splice to allow the DSL to control the process of converting the result of a splice into a Spliceable value. For example,
``ExampleDsl`... ${spliced_expr} ...` `` desugars to
```
$0splice0 = ExampleDsl::lift(spliced_expr);
ExampleDsl::makeTree( ... ${$0splice0} ...)
```
The use of the spliced temporary in the virtualized expression requires that the value has a subtype of `Spliceable` (which is more-or-less equivalent to requiring it to be an `ExampleDsl` expression). The `lift` function can then use an type that is a case type with where clauses to produce the `Spliceable` out of a variety of Hack types automatically. The example `lift` functions in the test cases do not currently test this capability.

Reviewed By: dlreeves

Differential Revision: D71738421

fbshipit-source-id: e5cb226e08066446c333dfbdde4beeee1abdc2ae