tolerate ill-formed Hack better
Summary:
## User-facing change

When there is bad Hack syntax in Hack notebook, the user should still be able to convert from a notebook to a .php file and back.

See [nb8_missing_semicolon.php](https://www.internalfb.com/diff/D72049793?permalink=2413523122343941) for an example.

We choose this for best user experience, since it's hard to give actionable errors when something goes wrong.

**note** source-controlled Hack notebooks have magic comments that show where the cells begin and end.

## How

Instead of treating Hack files as Hack, we treat them as a collection of start/end cell markers and a notebook-level metadata comment.

## Internal Behavior Changes

- The format changed slightly so that each cell has a beginning-of-cell and end-of-cell marker. The end-of-cell markers helps us do the following without having to try to be too clever (and without relying on parsing, which we're no longer doing when converting Hack->notebook):
   - exclude the `<?hh` at the top of the notebook
   - exclude the `async function gen_notebook_main_XXXX` function and its closing curly brace

- The error messages for bad metadata in magic comments got more accurate in a few places

## Guide for Reviewing

I recommend ignoring the code and looking at the tests. No one else has worked with this codebase and I had to change a lot.

Reviewed By: lindseyjin

Differential Revision: D72049793

fbshipit-source-id: f8310c0e43781a7ffc003b044f0fcb64b23c1b4e