Update SimpliHack attribute to take plain strings as well as promptlet classes for first argument
Summary:
Today, the SimpliHack attribute requires users to define a promptlet class and reference that as the first argument in the attribute.

Relaxing that to allow users to just pass in a simple string can make this experience lighter weight.

In this diff, I update the eval function in the simplihack_interpreter to allow for people to pass just a plain string in as an argument. If they pass in a promptlet class / class pointer - that is also still supported!

Reviewed By: dlreeves

Differential Revision: D72066507

fbshipit-source-id: 741919643247b6b436fe8468447a8cff2dc365ff