Use unstable simpli_hack feature test
Summary:
dlreeves noticed a problem where when he used the SimpliHack attribute, he correctly got the error that SimpliHack was an unstable feature. However, when he added the UnstableFeature attribute... he still got an error that SimpliHack was an unstable feature.

This was because I was inconsistent in my snake casing (sometimes using simplihack and sometimes using simpli_hack).

In this diff, I update the feature to be named 'simpli_hack' and added a test to show that when someone adds the UnstableFeature attribute for simpli_hack in the file- we have no errors!

Reviewed By: dlreeves

Differential Revision: D72001973

fbshipit-source-id: 51c08d77462eea9a1c75b5ce741bacbb3bc134d4